+++ src/main/java/edu/wpi/N/entities/employees/Laundry.java
@@ -1,4 +1,4 @@
-package edu.wpi.N.entities;
+package edu.wpi.N.entities.employees;
 
 public class Laundry extends Employee {
   public Laundry(int id, String name) {
