+++ src/main/java/edu/wpi/N/entities/employees/Translator.java
@@ -1,4 +1,4 @@
-package edu.wpi.N.entities;
+package edu.wpi.N.entities.employees;
 
 import java.util.LinkedList;
