+++ src/main/java/edu/wpi/N/views/AdminController.java
@@ -3,9 +3,9 @@ package edu.wpi.N.views;
 import edu.wpi.N.App;
 import edu.wpi.N.database.DBException;
 import edu.wpi.N.database.ServiceDB;
-import edu.wpi.N.entities.Employee;
-import edu.wpi.N.entities.Request;
-import edu.wpi.N.entities.Translator;
+import edu.wpi.N.entities.employees.Employee;
+import edu.wpi.N.entities.employees.Translator;
+import edu.wpi.N.entities.request.Request;
 import java.io.IOException;
 import java.net.URL;
 import java.util.GregorianCalendar;
