+++ src/main/java/edu/wpi/N/views/MapDisplayController.java
@@ -8,8 +8,8 @@ import edu.wpi.N.database.DBException;
 import edu.wpi.N.database.MapDB;
 import edu.wpi.N.database.ServiceDB;
 import edu.wpi.N.entities.DbNode;
-import edu.wpi.N.entities.Doctor;
 import edu.wpi.N.entities.Path;
+import edu.wpi.N.entities.employees.Doctor;
 import edu.wpi.N.qrcontrol.QRGenerator;
 import java.io.IOException;
 import java.util.ArrayList;
